Datagram test fixups
- debug test counters are now static Connection member variables
- tests can count on server and client side
- moved messaging containers to messages.{hpp,cpp}
